pyproject.toml
[project]
name = "microfoon"
version = "0.1.0"
description = "USB Audio Processor with Gemini AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai",
    "watchdog",
    "pydub",
    "sqlalchemy",
    "rich",
    "python-dotenv",
    "audioop-lts; python_version >= '3.13'",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["microfoon*"]
exclude = ["recordings*", "test_usb_drive*", "venv*"]

